+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = querymagic
-version = 0.3.21
+version = 0.3.23
 author = avinesh18
 author_email = avinesh18@github.com
 description = Query Magic
