+++ src/querymagic.egg-info/PKG-INFO
@@ -1,6 +1,6 @@
 Metadata-Version: 2.1
 Name: querymagic
-Version: 0.3.21
+Version: 0.3.23
 Summary: Query Magic
 Home-page: https://github.com/Avinesh18/lse-python-packages
 Author: avinesh18
